refactor(observability): consolidate notifications under observability/ (#164)
Phase 8b of the src/srunx/ cleanup (#156):

Moves:
  - src/srunx/notifications/         → src/srunx/observability/notifications/
  - src/srunx/formatters.py          → src/srunx/observability/notifications/formatting.py
  - SlackCallback class body         → src/srunx/observability/notifications/legacy_slack.py

Renames:
  - SlackWebhookDeliveryAdapter      → SlackWebhookAdapter
    (package name already scopes it; ``Delivery`` was redundant)

Shims preserve legacy import paths during the Phase 8 transition:
  - ``srunx.notifications.*`` aliases every submodule via sys.modules so
    ``from srunx.notifications.service import ...`` (and monkey-patches)
    keep working identically.
  - ``srunx.formatters`` re-exports ``SlackTableFormatter`` /
    ``SlackNotificationFormatter`` from the canonical path.
  - ``srunx.callbacks`` resolves ``SlackCallback`` lazily via PEP 562
    ``__getattr__`` (with a ``TYPE_CHECKING`` re-export for static
    analysis) so ``from srunx.callbacks import SlackCallback`` keeps
    working without the cyclic import between the ``Callback`` base
    and its subclass.

Test updates:
  - ``tests/test_callbacks.py`` now patches
    ``srunx.observability.notifications.legacy_slack.WebhookClient``
    (the canonical home of ``SlackCallback``'s ``WebhookClient`` binding).
  - All other test files keep importing via the legacy shim paths —
    exercise the shim at the same time.

Behavior unchanged — no API changes beyond the class rename (and the
rename is source-compatible for existing callers because the new class
is imported by the same module).
