+++ src/srunx/observability/notifications/adapters/base.py
