+++ src/srunx/observability/notifications/presets.py
