+++ src/srunx/observability/notifications/service.py
@@ -18,7 +18,7 @@ from srunx.db.repositories.endpoints import EndpointRepository
 from srunx.db.repositories.events import EventRepository
 from srunx.db.repositories.subscriptions import SubscriptionRepository
 from srunx.db.repositories.watches import WatchRepository
-from srunx.notifications.presets import should_deliver
+from srunx.observability.notifications.presets import should_deliver
 
 
 class NotificationService:
