+++ tests/integration/test_end_to_end_notification.py
@@ -17,6 +17,8 @@ from typing import Any
 
 import anyio
 import pytest
+from srunx.notifications.adapters.base import DeliveryAdapter, DeliveryError
+from srunx.notifications.service import NotificationService
 
 from srunx.db.connection import open_connection
 from srunx.db.migrations import apply_migrations
@@ -26,13 +28,11 @@ from srunx.db.repositories.endpoints import EndpointRepository
 from srunx.db.repositories.events import EventRepository
 from srunx.db.repositories.subscriptions import SubscriptionRepository
 from srunx.db.repositories.watches import WatchRepository
-from srunx.notifications.adapters.base import DeliveryAdapter, DeliveryError
-from srunx.notifications.service import NotificationService
 from srunx.pollers.delivery_poller import DeliveryPoller
 
 
 class RecordingAdapter:
-    """Stand-in for ``SlackWebhookDeliveryAdapter`` that records calls."""
+    """Stand-in for ``SlackWebhookAdapter`` that records calls."""
 
     kind = "slack_webhook"
